chore: move context manager from util to common
